Implement new fcc option.
"Put a copy of the sent message in the same folder as the message being replied
to, unless the message being replied to is in the inbox -- in that case it will
go into the sent folder".

This is mainly a self-service option but it helps me keep my folders tidy.